Clean up and remove unused components
