+++ frontend/src/components/StatementForms/StatementForm.tsx
@@ -1,30 +1,15 @@
-import {
-  Box,
-  Container,
-  Grid,
-  GridItem,
-  Heading,
-  Input,
-  Stack,
-  StackDivider,
-  Table,
-  Tbody,
-  Td,
-  Text,
-  Th,
-  Thead,
-  Tr
-} from "@chakra-ui/react";
+import { Stack } from "@chakra-ui/react";
 import { AuthContext } from "contexts/AuthContext";
 import { useLocales } from "hooks/useLocales";
 import { FC, useCallback, useContext, useState } from "react";
 import { useForm } from "react-hook-form";
-import { IClientStatementDto, IStatementFieldInputDto } from "services/backend/nswagts";
+import { IClientStatementDto } from "services/backend/nswagts";
 
-import InputDKK from "./InputDKK";
-import StatementHeaderRow from "./StatementHeaderRow";
-import StatementRow from "./StatementRow";
 import StatementSection from "./StatementSection";
+import StatementSectionTable from "./StatementSectionTable";
+import StatementTableRow from "./StatementTableRow";
+import StatementTableSubHeading from "./StatementTableSubHeading";
+import StatementTableSubHeadings from "./StatementTableSubHeadings";
 
 interface Props {
   statement: IClientStatementDto;
@@ -45,6 +30,122 @@ const StatementForm: FC<Props> = ({ statement }) => {
     <form onSubmit={onSubmit}>
       <Stack>
         <StatementSection heading={t("statements.section1.heading")}>
+          <StatementSectionTable>
+            <StatementTableRow text={t("statements.section1.mushrooms")} tax="0.25" />
+            <StatementTableRow text={t("statements.section1.tomatoCucumberHerbs")} tax="2.00" />
+            <StatementTableSubHeadings h2={t("statements.expences")} />
+            <StatementTableRow
+              text={t("statements.boughtPlants")}
+              subText={t("statements.section1.boughtPlantsDesc")}
+              tax="2.00"
+              tooltip="Tooltip her"
+            />
+          </StatementSectionTable>
+        </StatementSection>
+        <StatementSection heading={t("statements.section3.heading")}>
+          <StatementSectionTable>
+            <StatementTableRow text={t("statements.section3.carrot")} tax="3.00" />
+            <StatementTableRow text={t("statements.section3.pea")} tax="3.00" />
+            <StatementTableSubHeadings h2={t("statements.expences")} />
+            <StatementTableRow
+              text={t("statements.boughtPlants")}
+              subText={t("statements.section3.boughtPlantsDesc")}
+              tax="3.00"
+            />
+          </StatementSectionTable>
+        </StatementSection>
+        <StatementSection heading={t("statements.section4.heading")}>
+          <StatementSectionTable>
+            <StatementTableRow text={t("statements.section4.onions")} tax="1.60" />
+            <StatementTableRow text={t("statements.section4.plants")} tax="1.60" />
+            <StatementTableRow text={t("statements.section4.flowers")} tax="1.60" />
+            <StatementTableSubHeadings h2={t("statements.expences")} />
+            <StatementTableRow
+              text={t("statements.boughtPlants")}
+              subText={t("statements.section3.boughtPlantsDesc")}
+              tax="1.60"
+            />
+          </StatementSectionTable>
+        </StatementSection>
+        <StatementSection heading={t("statements.section7.heading")}>
+          <StatementSectionTable>
+            <StatementTableRow text={t("statements.section7.plants")} tax="4.50" />
+            <StatementTableSubHeadings h2={t("statements.expences")} />
+            <StatementTableRow text={t("statements.boughtPlants")} tax="4.50" />
+          </StatementSectionTable>
+        </StatementSection>
+        <StatementSection heading={t("statements.section8.heading")}>
+          <StatementSectionTable h1="" h2="" h3="">
+            <StatementTableSubHeading>
+              {t("statements.section8.subHeading1")}
+            </StatementTableSubHeading>
+            <StatementTableSubHeadings
+              h2={t("statements.turnoverExlMoms")}
+              h3={t("statements.taxIs")}
+            />
+            <StatementTableRow text={t("statements.section8.applesPearsOther")} tax="5.00" />
+            <StatementTableSubHeadings h2={t("statements.expences")} />
+            <StatementTableRow
+              text={t("statements.section8.packagingCost")}
+              subText={t("statements.section8.packagingCostDesc")}
+              tax="5.00"
+            />
+            <StatementTableSubHeading>
+              {t("statements.section8.subHeading2")}
+            </StatementTableSubHeading>
+            <StatementTableSubHeadings
+              h2={t("statements.turnoverExlMoms")}
+              h3={t("statements.taxIs")}
+            />
+            <StatementTableRow text={t("statements.section8.cherry")} tax="4.65" />
+            <StatementTableRow text={t("statements.section8.plum")} tax="4.65" />
+            <StatementTableRow text={t("statements.other")} tax="4.65" />
+            <StatementTableSubHeading>
+              {t("statements.section8.subHeading3")}
+            </StatementTableSubHeading>
+            <StatementTableSubHeadings
+              h2={t("statements.turnoverExlMoms")}
+              h3={t("statements.taxIs")}
+            />
+            <StatementTableRow text={t("statements.section8.currant")} tax="4.60" />
+            <StatementTableRow text={t("statements.section8.strawberry")} tax="4.60" />
+            <StatementTableRow text={t("statements.other")} tax="4.60" />
+          </StatementSectionTable>
+        </StatementSection>
+      </Stack>
+    </form>
+  );
+};
+export default StatementForm;
+/*
+<Table variant="unstyled">
+            <Thead>
+              <Tr>
+                <Th></Th>
+                <Th w="30%">{t("statements.turnoverExlMoms")}</Th>
+                <Th w="15%">{t("statements.taxIs")}</Th>
+              </Tr>
+            </Thead>
+            <Tbody>
+              <StatementTableRow text={t("statements.section1.mushrooms")} tax="0.25" />
+              <StatementTableRow text={t("statements.section1.tomatoCucumberHerbs")} tax="2.00" />
+              <StatementTableSubHeadings h2={t("statements.expences")} />
+              <StatementTableRow
+                text={t("statements.boughtPlants")}
+                subText={t("statements.section1.boughtPlantsDesc")}
+                tax="2.00"
+                tooltip="Tooltip her"
+              />
+            </Tbody>
+          </Table>
+
+
+
+
+
+
+
+<StatementSection heading={t("statements.section1.heading")}>
           <StatementHeaderRow h1={t("statements.turnoverExlMoms")} h2={t("statements.taxIs")} />
           <StatementRow text={t("statements.section1.mushrooms")} tax="0.25" />
           <StatementRow text={t("statements.section1.tomatoCucumberHerbs")} tax="2.00" />
@@ -92,11 +193,8 @@ const StatementForm: FC<Props> = ({ statement }) => {
           <StatementRow text={t("statements.section7.plants")} tax="4.50" />
           <StatementRow text={t("statements.boughtPlants")} tax="4.50" />
         </StatementSection>
-      </Stack>
-    </form>
-  );
-};
-export default StatementForm;
+*/
+
 /*
 <Stack>
         <StatementSection2 heading={t("statements.section1.heading")}>
