--- frontend/src/components/StatementForms/StatementHeaderRow.tsx
@@ -1,40 +1,0 @@
-import {
-  Button,
-  Flex,
-  Grid,
-  Heading,
-  HStack,
-  IconButton,
-  Input,
-  Stack,
-  Text
-} from "@chakra-ui/react";
-import { AuthContext } from "contexts/AuthContext";
-import { setAuthToken } from "hooks/useAuth";
-import { useLocales } from "hooks/useLocales";
-import { useRouter } from "next/router";
-import { FC, useCallback, useContext, useEffect, useState } from "react";
-
-interface Props {
-  h1: string;
-  h2: string;
-}
-
-const StatementHeaderRow2: FC<Props> = ({ h1, h2 }) => {
-  const { t } = useLocales();
-  const router = useRouter();
-  const { activeUser } = useContext(AuthContext);
-
-  return (
-    <>
-      <Text></Text>
-      <Text textTransform="uppercase" fontWeight="bold" textColor="gray.600" fontSize="sm">
-        {h1}
-      </Text>
-      <Text textTransform="uppercase" fontWeight="bold" textColor="gray.600" fontSize="sm">
-        {h2}
-      </Text>
-    </>
-  );
-};
-export default StatementHeaderRow2;
